requirements.txt
discord.py
ffmpeg
yt-dlp
validators
youtube-search
PyNaCl
boto